IMPORTANT: commit before taking away some now useless code parts + new code on interpolation foot making.
New code: creation of foots for the data that are 2D gaussians in 3D. While the code for
making the feet run, the resutls of it are not yet used as the new data gotten is still
in astropy.coordinates.SkyCoord cartesian carrington heliographic format (as opposed in
a np.ndarray 4D cube). Still need to decide if I need to change the main way I save the
data, maybe in a SkyCoord object, we'll see.

Old code: This commit way mainly done to tell the reader that some old code will be deleted
in the next commit. This code is mainly on the 'datatype' argument options as I believe
only the 'raw' and maybe 'rawContour' options should be kept as they moslty give similar
(sometimes better) results than the other options while the data doesn't need preprocessing,
i.e. it is easier to make sense of the choice maybe in the data used.