MAJOR: 99 max char limit + Interpolation angles
A lot of files have changed bcs of reformating some codes so that there is 99 char max per code line. Now creating the interpolated curve using the polynomial parameters directly and ending the curve right at the surface of the sun or ending it a little further than the borders of the cube (maybe change it to a recalculated border given the max index.). Also added the angle values compared to the image plane with a colorbar in the plot. Cleaned up the Projection.polynomial_projectiom code and made the plot choices more modulable.