MIDDLE: restructuring and hdf5 test data added.
I moved the polynomial fit creation class for the HDF5 data file to the get_interpolation.py file. Created (and commented) a small code to imshow() the polar reprensentation of the full image gotten from SDO. Not error seen in the choosing of the Sun's center. Added a new code file (cubes_add_test.py) where I used the hdf5 file and corresponding creation class to add a new group to the file with test data. For know, only the Sun's surface was added.