MAJOR: add dataclasses + code more modulable (data + fake data). 3D visualisation ~ works - WORK IN PROGRESS.
A lot of changes. Added dataclasses to make the code more readible and understandable. Also, added more parent classes so that the code is more modulable. Its to help when I need to visualise the real and/or fake data. For now, the hdf5 creation for the real data most likely doesn't work, only for the fake data. The 3D animation kind of works for the fake data, but still has major flaws. The animation for the real data should still work properly. Still a lot to change and a lot of new errors must have appeared.