MAJOR: structure changes to make code smaller and more readible.
Updated the animation dataclasses so that the 3D information is easily accessible and the structure is understandable. Furthermore, for the 3D animation, there shouldn't be any more problems on if the xt_min is the same for all the data set as each 3D coordinates has its own translation argument in the animation. Now the plot is centered on the Sun. The code still doesn't work as I also need to update and change the structure of the HDF5 file -- WORK IN PROGRESS.