MIDDLE: separation of processing and result plotting. Can choose to not multiprocess.
The creation of the re-projection data and the actual plotting of the data was separated where the plotting becomes a child class of the processing class. This was done to be able to use the result in the re-projection in another class to then recreate FITS files for running Dr. Auchere's new_toto.pro code on it. That being said, as of now, this plan was stopped as I don't think it can give me any usefull data for debugging. Will delete the useless classes later.