NOTICE - Last commit message is wrong.
Did a git commit -am as opposed to -m so the information in the last commit is not complete. In the last commit, I also added a Torus to the fake data and a 3D cross inside the fake cube data. Furthermore, the file data was added in this commit is useless for now and most likely will always be useless. Surely I am going to remove it in no time.