MINOR: started adding code to fusion hdf% files. gitignore updated and addedd to tracked files.
writting a code to fusion the fake and real data files but need to commit first as the #d visualisation is more important for now. Also added the .gitignore file as I had forgotten to do so and updated the common module so that it contains an batch job template for when running my code on the server using slurm.