MINOR: updated common package and added some test to print the satellite positions.
I have an error somewhere in the satellite positions so added some separate bits of code to print the positions used in my codes. Also updated the common package by using @overload in 2 methods. Need to check the satellite positions direclty in the server as I have a weird result problem.