MAJOR: found the problem in the visualisation and corrected it
The problem in k3d comes from the library itself. The first cubes is plotted correctly but weirdly, for the next ones, the reference point for the cube itself changes meaning that if the cube size changes over time, then the positions of the rest of the points is wrong. Therefore I forced the cube shapes for all the cubes to be the max possible value given the chosen cubes for the specific run. Now the position of the cubes relative to the sun and to the satellite seems to be right (checked the line of sight data.).