MAJOR: Works with the fusion data (i.e. real + fake datafile).
Updated the code so that it works on the fusion datafile. The dataclasses were also updated so that the correct fake data index can be chosen and the name of the datagroups has been improved. Also, some @overloads were added as some methods have changed. Code runs well except when not setting the 'sdo_pov' or 'stereo_pov' arguments to True. Other tests might also fail but all the ones I have done work.