MINOR: add an intersection method to test
I added an intersection method using jax. That being said, the method is wrong for now but keeping it to test if it works well. Also added a TODO to the __init__ file as there is a new test that I should try to see if the perspective is taken into account in the new_toto.pro code.