MIDDLE: projection code now works with the fusion data (at least partially, more changes need to be made).
I change the projection code so that it can take the real data or the fusion data as input. Also implemented the config.yml file in the code to get the datafile paths. While the code works for some arguments, it still doesn't work for others, hence there is still a good amount of changes to be done.