MIDDLE: path changes and corresponding updates in some of the codes.
I imported some of the data that I had on the server to properly update the path setup in the config.yml. I also deleted a lot of data that I might as well recreate if I ever do need it again. The config file and some codes were updated given the new changes in the paths.