MIDDLE: projection update so that it works with the fusion hdf5 datafile.
I updated the projection code so that it works with the fusion data and the test data is plotted properly. The code was also cleaned up to be a little more readible and the colours for the plot was changed. The code runs properly but I still have problems in the positionning of the testing/fake dataset.