MAJOR: test cube added. Visualisation arguments update. Update in the usage of cube shape. Voxel colours added. Config file usage update. Docstrings updated.
I changed the visualisation code so that it is more readible. Updated the docstring, added the possibility to easily change the voxel colours. Change the usage of the max shape so that only the max shape of each datatype is used for the visualisation. Added the possibility to also visualise the test data (with the fake and real one). Major update on how to decide what to plot. The high number of possibilities made me choose a list[str] as a single argument when choosing the pov and data visualisation choices. For now, the code seems to work properly.