MINOR: add .pyi files. Type annotation improvements.
I added .pyi file to the common library and to the .gitignore of this git repository. Also added the final decorator to some methods. Wanted to add .pyi file to this repo, but the problem is that pylance is trippin, if it exists, and only checks the .pyi files for function definitions. Hence, if I want to use .pyi files, I need to name all the different methods, and if in another code I check the import method definition, I won't get the actual function definition, but only the type annotations for that function as it sends me to the .pyi file. Completly retarded behavior if my opinion has any value.