MAJOR: base class created and started emplementing the new fit envelope -- WORK IN PROGRESS.
Created a base class so that the same code can be used in multiple other classes as new code is being added (i.e. the new fit envelope processing). Started testing the fit envelope in the sdo_projection file, but it is still in the testing mode, and it doesn't work yet.