MAJOR: fit and corresponding envelope processing done properly.
I updated the code so that now the fit reprojection points are uniformly spaced (along the curve) in 2D. Furthemore, the envelope processing was also updated so that the envelope stops at the sun disk as seen from SDO. Moreover, the corresponding points representing the envelope are also now uniformly spaced along the curve. The sdo reprojection code seems to work properly with the modifications added.