MIDDLE: sdo reprojection update as the fit and envelope processing changed. Type:ignore added where relevant.
The sdo reprojection code changed a little bit as the fit and envelope processing was just updated. As such, two new dataclasses where added in the corresponding code file so that the new processed data is nicely formated and henced accessible. Furthermore, a lot of type:ignore comments were addeed throughout the code where no other solution was found (when fetching data with a path string inside an HDF5 file, the IDE doesn't know if the path points to a group or a dataset, hence the pylance errors.). Codes seems to run properly.