MIDDLE: can now plot multiple integration times and polynomial fit orders at the same time.
Improved the code so you can now plot multiple integration time contours with the corresponding fit and fit envelopes. Furthermore, you can also choose multiple polynomial fit orders to be plotted at the same time. Finally, the file fetching from the server was also improved so that if there is an error in the fetching of the file, then the code tries again (given a sleep time) till a set number of retries.