MAJOR: sdo image warping added. Doens't work properly yet -- WORK IN PROGRESS.
I added the codes necessary for the warping of the sdo image in between the fit envelopes. While the code does run, the results are still wrong. Furthermore, the adding of the warping has made the code less readible and more error prone. Extensive clean up is needed after making the warping work. -- WORK IN PROGRESS.