MIDDLE: fixed the SDO image warp. Small code clean ups.
The SDO image warping now works properly. Also I added type:ignore comments when using values from a complex dictionary. Cleaned up some of the code so that the argument choices make a little more sense. Next step would be to update the axes of the warped image and thnik about an improvement to the envelope fit as the form of the envelope fit next to the feet positions messes up the corresponding warped image.