MINOR: Bug fixes for the full data integration -- STILL DOESN'T WORK.
I updated bits of code that created problems when trying to do the fit of the full integration of the data. The code still doesn't work but haveen't been able to understand where the problem comes from as only the fit parameters of one axe are all zeros. Also cleaned up the code a little, added config file value usage and some type:ignore comments.