MINOR: adding full integration to sdo projection.
I started changing codes to incorporate the full integration with the corresponding fit inside the sdo projection code. Some small improvements in the code for a better use in the sdo projection code. Also, swapped some type:ignore comments to using typing.cast. Main of the changes are not yet commited as they are not finnished.