MIDDLE: dataclasses moved around due to circular imports.
I have started to move around some of the main dataclasses used for the projection of the data to try and not have circular imports. I am still working on the update for the full integration data but stuck because of circular imports. A major change is to be done and most of the changes done in my code hasn't yet been commited.