MAJOR: full integration implementation and code improvements.
The full integration of the no duplicate and all data were added to the reprojection and the corresponding warped image can also be created. The code runs but there seems to be an error somewhere on if the right data is chosen as big changes in the reprojected fit are sometimes seen (the fit should move a little but not by that much, at least not the feet positions). Furhtermore, a lot of code clean up is needed to delete the deprecated comments and update/add new doctrings.