MIDDLE: arc-length annotations added. Corresponding update in the common package.
I added the annotations of the arc-length on the create fit envelopes. Code still needs changes if the arclength needs to be added to Auchere's envelope and the fit themselves. The arc-length annotation comes from the new code added in the now updated common package. The angles for the arc-length text annotations are still wrong. Except these points, the code seems to be working properly.