MAJOR: 'all sdo images' option add. Warped 'integration' add.
I added the option to be able to choose to plot all the sdo images, even those for which there is no 3D data. Also, started creating the final plot in Auchere's paper. The problem being that I wasn't able to properly test it as sometimes there is a connection error with the server. Need to test it directly on the server itself. Aslo, mostlikely need to change the cube creation so that I can have the time integration even for dates where there is no data.