MIDDLE: type annotation improvements. Added all dates to the integration data. NEEDS TESTING ON SERVER.
I updated the type annotations of multiple codes given that I have just gotten knew insights on the typing module. Furthermore, I changed the hdf5 datafile creation so that all the different cube indexes are added when doing the time integration. This is so that for the final projection of the datasets, I can also see the integration data for the dates in which there is no actual volumetric data.