MIDDLE: 'all dates' option added to the k3d visualisation. Type annotation improvements.
I added the possibility to visualise all the dates in the k3d visualisation (even the dates where there is no data). The code seems to work properly. Can't say for sure as there is an error in the HDF5 data (must have made a mistake in the indexes when adding 'all dates' to the integration data). Also, improved some of the type annotations and took away all the type:ignore comments in the animation_code code.