MIDDLE: cube indexes for all cubes changed to 'Time indexes'.
I changed the time indexes for all the cubes so that they match the indexes of 'Time indexes'. Makes the following usage of the data way easier and understandable (not sure why I hadn't done so earlier). The code still needs testing on the server and the other Python codes using the HDF5 file data still haven't been updated correspondly.