MINOR: updated indexes in animation_code.py.
As I had just changed the cube time indexes in the HDF5 data so that they always coincide with the 'Time indexes' values, I also correspondingly updated animation_code. While the code does run, I cannot make sure that the update was done properly as the update in the HDF5 file is crealy wrong. Looking for the error right now.