MAJOR: adding new dataclasses for the creation of the final warp integration plot. WORK IN PROGRESS.
I started adding new dataclasses so that I can easily access the needed warp integration data and metadata. The work is still not finished as I am not yet sure if I should change the main structure of the sdo_reprojection code or of the corresponding dataclasses as right now the addition of the warp dataclasses is a fuck fest. Will most likely need to change some part of sdo_reprojection and minor changes to two or three dataclasses.