MIDDLE: 'plotting' directory creation. sdo_reprojection and warp_information codes updates. WORK IN PROGRESS.
I starting separating the plotting codes from the sdo_reprojection processing to make the code more readible. Started changing the sdo_reprojection and the warped_information code to be able to get the necessary data for the warp integration plot. Started creating the plotting code for the final warp integration figure.