MAJOR: warp related code updates and plotting function 'finished'. Relative imports error fixed. TESTING FAILED, NEED TO FIND THE ERROR.
I updated the warp_information class as the corresponding dataclass was also updated. I also finishde the code for the warp integration plot and the warp integration data pickling (from the multiprocessing manager queue) and the restructuration of the data for the ease of use in the warp integration plotting code. That being said, while trying to run the code, there is an error saying that the warp data is set to None. Needs further debugging.