MINOR: warp image shape depends on cumulative distance nb of points.
I took away the predifined warp image shape so that it now depends directly on the number of points chosen for the cumulative distance and hence on the number of points in the fits and envelope. This was done so that for each pixel, you have a corresponding cumulative distance value and an angle.