MINOR: nearest neighbour error fix. Added test plot.
I fixed the error in which the nearest neighbour (to the get the corresponding angle value) wasn't the right one. This was due to the fact that I hadn't normalised the radial distance and polar angle values (so that the units don't change the relative weights each position has). Also, I added a test plot to be able to visualise the result and see if the code seems to be working properly.