MINOR: Fix in the coordinate normalisation and docstring add/update.
I fixed the 'normalisation' of the coordinates when computing the nearest neighbours (to get the corresponding fit angle). Now the normalisation between 0 and 1 is done on the fit coordinates and the same 'borders' are used for the envelope middle path (so that the distance between the two makes sense). Furthermore, I had added docstring to the __init__ of the class and updated another doctring.