MIDDLE: cubes creation code with acquisition cadence of 1 minute finished - NEEDS TESTING ON THE SERVER.
I just finished updating the HDF5 data creation code where the new cadence of 1 minute for the SDO acquisitions was taken into account. Wasn't able to test the code as it seems that the archive in the IAS server is partially down and the URL for fetching SDO metadata is also given error:500 (the website is also hosted on said server).