MIDDLE: correction and improvements in the sdo metadata filtering. Added 'IAS path' to HDF5 data - HDF5 CODE NEEDS TESTING ON SERVER.
I corrected the all_sdo_dates.py code as I was filtering the SDO metadata (for a cadence of 1 minute) wrongly. There is still 20 images that are missing, will need to check later. Also, I updated the code so that I am using .replace method of datetime.datetime objects as opposed of my own datetime comparaison method. Furthermore, I updated the HDF5 data creation code so that the IAS server paths to the corresponding SDO images is also directly saved inside it. While the filtering code seems to work, I still need to test the HDF5 creation code on the server.