MINOR: sdo_reprojection code uses 'IAS paths' from HDF5 file, change values to instance attributes, corresponding code clean up and docstring update.
I updated the sdo_reprojection code so that it uses the 'IAS paths' dataset from the HDF5 file. The now deprecated method to find the timestamps was therefore taken away. I also changed some values used in the warping of the SDO acquisitions to instance attributes and updated the corresponding docstrings. I also cleaned up a little bit some other codes and updated a dataclass so that it can also contain the needed ias_path variable.