MINOR: sdo metadata datetime microseconds set to 0). - WORK IN PROGRESS.
I changed the processing of the SDO metadata so that all the datetimes gotten from my processing have their microseconds set to 0. Still a work in progress as I have problems with the difference between the STEREO B dates (gotten from the PNG filenames) and the corresponding SDO dates. Committing as the changes were done on the server, so if I forget tomorrow to log on the server first, I won't have commit clashes.