MIDDLE: HDF5 data code update. Seems to work properly.
I fixed the errors in the HDF5 data creation code. Properly converted the STEREO dates to corresponding SDO dates. The code seems to run properly on the server. I should update it so that it is more efficient and so that it uses way less RAM. That being said, to do so, I would need to pretty much change the whole structure of said code and add some new methods for when using and appending values inside an HDF5 file. No time for now.