NULL
Nothing really changed. Just added some prints and took away some others. Only committing to run the code on the server as right now there seems to be some problems related to the connection with the server. The running time is basically only the wait time for the data to be fetched even though, when the connection works 'properly' it shouldn't take that long as there is more processes than concurrent fetches so the downloading is supposed to be continuous (it isn't right now).