MINOR: update in the serialisation of the data when using manager.Queue() from the multiprocessing module.
I had to update the serialisation process as I didn't know that when using a multiprocessing queue, the data is serialised twice (once to access the shared memory, once to leave it after everything is finished). Hence, I added an instance attribute letting me know if the data was already serialised once (as a preprocessing step is done to save RAM).